requirements.txt
alabaster==0.7.10
Babel==2.6.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
Django==2.0.5
django-bootstrap4==0.0.6
django-braces==1.13.0
docutils==0.14
Faker==0.8.16
idna==2.7
imagesize==1.0.0
Jinja2==2.10
MarkupSafe==1.0
misaka==2.1.0
packaging==17.1
Pillow==5.1.0
pycparser==2.18
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
requests==2.19.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.5
sphinxcontrib-websupport==1.0.1
text-unidecode==1.2
urllib3==1.23
